ai/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
transformers==4.43.3
torch==2.3.1
numpy==1.26.4
pydantic==2.8.2
python-dotenv==1.0.1
